% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/some_function.R
\name{someFun}
\alias{someFun}
\title{Some function}
\usage{
someFun(x, y)
}
\arguments{
\item{x, y}{two values to be summed}
}
\value{
the sum x+y
}
\description{
Some function
}
\examples{
someFun(2,3)
}

